+++ src/fastfetch.c
@@ -1226,8 +1226,8 @@ static void parseOption(FFinstance* instance, FFdata* data, const char* key, con
             optionParseString(key, value, &instance->config.libOSMesa);
         else if(strcasecmp(subkey, "-opencl") == 0)
             optionParseString(key, value, &instance->config.libOpenCL);
-        else if(strcasecmp(subkey, "-cjson") == 0)
-            optionParseString(key, value, &instance->config.libcJSON);
+        else if(strcasecmp(subkey, "-jsonc") == 0)
+            optionParseString(key, value, &instance->config.libJSONC);
         else if(strcasecmp(subkey, "-wlanapi") == 0)
             optionParseString(key, value, &instance->config.libwlanapi);
         else if(strcasecmp(key, "-pulse") == 0)